Refactored module classes into data package
