+++ src/main/java/duke/Duke.java
@@ -9,8 +9,8 @@ import duke.exceptions.ModBadRequestStatus;
 import duke.exceptions.ModException;
 import duke.exceptions.ModFailedJsonException;
 import duke.exceptions.ModTimeIntervalTooCloseException;
-import duke.modules.ModuleInfoDetailed;
-import duke.modules.ModuleInfoSummary;
+import duke.modules.data.ModuleInfoDetailed;
+import duke.modules.data.ModuleInfoSummary;
 import duke.util.JsonWrapper;
 import duke.util.ParserWrapper;
 import duke.util.Storage;
