+++ src/main/java/duke/modules/data/Attributes.java
@@ -1,4 +1,4 @@
-package duke.modules;
+package duke.modules.data;
 
 public class Attributes {
