+++ src/main/java/duke/modules/data/ModuleInfoDetailed.java
@@ -1,6 +1,4 @@
-package duke.modules;
-
-import java.util.List;
+package duke.modules.data;
 
 public class ModuleInfoDetailed {
