+++ src/main/java/duke/modules/data/ModuleInfoSummary.java
@@ -1,4 +1,4 @@
-package duke.modules;
+package duke.modules.data;
 
 import java.util.Arrays;
